# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: Creative Sports League API
  version: "1.0.0"
  description: REST API for the Creative Sports League platform — leagues, challenges, submissions, scoring, and sponsor management.
servers:
  - url: /api/v1
paths:
...
  /leagues/{leagueId}/enrollments:
    post:
      summary: Enroll a participant in a league
      operationId: enrollParticipant
      parameters:
        - $ref: "#/components/parameters/leagueId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [participantId]
              properties:
                participantId:
                  type: string
                  minLength: 1
      responses:
        "200":
          description: Participant enrolled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /challenges/{challengeId}/submissions:
    get:
      summary: List submissions for a challenge (paginated)
      operationId: listSubmissions
      parameters:
        - $ref: "#/components/parameters/challengeId"
        - name: page
          in: query
          description: Page number (≥ 1, default 1)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Items per page (1–100, default 20)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Paginated submissions
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    enum: [true]
                  data:
                    type: array
                    items:
                      type: object
                  meta:
                    $ref: "#/components/schemas/OffsetMeta"
        "500":
          $ref: "#/components/responses/ServerError"
    post:
      summary: Submit an entry to a challenge
      operationId: submitEntry
      parameters:
        - $ref: "#/components/parameters/challengeId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [participantId, artifact]
              properties:
                participantId:
                  type: string
                  minLength: 1
                artifact:
                  $ref: "#/components/schemas/SubmissionArtifact"
                isPublic:
                  type: boolean
      responses:
        "201":
          description: Submission created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/ValidationError"
  /challenges/{challengeId}/leaderboard:
    get:
      summary: Leaderboard for a challenge (paginated)
      operationId: getLeaderboard
      parameters:
        - $ref: "#/components/parameters/challengeId"
        - name: page
          in: query
          description: Page number (≥ 1, default 1)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Items per page (1–100, default 20)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Paginated leaderboard
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    enum: [true]
                  data:
                    type: array
                    items:
                      type: object
                  meta:
                    $ref: "#/components/schemas/OffsetMeta"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /submissions/{submissionId}/score:
    post:
      summary: Score a submission
      operationId: scoreSubmission
      parameters:
        - name: submissionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [judgeId, criteriaScores, rationale]
              properties:
                judgeId:
                  type: string
                  minLength: 1
                criteriaScores:
                  type: array
                  items:
                    $ref: "#/components/schemas/CriteriaScore"
                rationale:
                  type: string
                  minLength: 1
      responses:
        "200":
          description: Submission scored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/ValidationError"

  # ── Sponsors ───────────────────────────────────────────────────────────────
...
components:
  parameters:
    leagueId:
      name: leagueId
      in: path
      required: true
      schema:
        type: string
    challengeId:
      name: challengeId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
          enum: [true]
        data:
          type: object
    OkCreated:
      type: object
      properties:
        ok:
          type: boolean
          enum: [true]
        data:
          type: object
    ErrorResponse:
      type: object
      properties:
        ok:
          type: boolean
          enum: [false]
        error:
          type: string
    OffsetMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    CriteriaScore:
      type: object
      required: [criteriaName, score]
      properties:
        criteriaName:
          type: string
          minLength: 1
        score:
          type: number
          minimum: 0
          maximum: 100
    SubmissionArtifact:
      type: object
      required: [url]
      properties:
        url:
          type: string
          minLength: 1
        mimeType:
          type: string
        description:
          type: string
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Unprocessable entity — validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"